+++ x/auth/simulation/decoder.go
@@ -19,11 +19,13 @@ func DecodeStore(cdc *codec.Codec, kvA, kvB tmkv.Pair) string {
 		cdc.MustUnmarshalBinaryBare(kvA.Value, &accA)
 		cdc.MustUnmarshalBinaryBare(kvB.Value, &accB)
 		return fmt.Sprintf("%v\n%v", accA, accB)
+
 	case bytes.Equal(kvA.Key, types.GlobalAccountNumberKey):
 		var globalAccNumberA, globalAccNumberB uint64
 		cdc.MustUnmarshalBinaryLengthPrefixed(kvA.Value, &globalAccNumberA)
 		cdc.MustUnmarshalBinaryLengthPrefixed(kvB.Value, &globalAccNumberB)
 		return fmt.Sprintf("GlobalAccNumberA: %d\nGlobalAccNumberB: %d", globalAccNumberA, globalAccNumberB)
+
 	default:
 		panic(fmt.Sprintf("invalid account key %X", kvA.Key))
 	}