+++ x/auth/simulation/genesis.go
@@ -98,7 +98,7 @@ func RandomizedGenState(simState *module.SimulationState) {
 func RandomGenesisAccounts(simState *module.SimulationState) (genesisAccs exported.GenesisAccounts) {
 	for i, acc := range simState.Accounts {
 		bacc := types.NewBaseAccountWithAddress(acc.Address)
-		var gacc exported.GenesisAccount = &bacc
+		var gacc exported.GenesisAccount = bacc
 
 		// Only consider making a vesting account once the initial bonded validator
 		// set is exhausted due to needing to track DelegatedVesting.
@@ -115,7 +115,7 @@ func RandomGenesisAccounts(simState *module.SimulationState) (genesisAccs export
 				endTime = int64(simulation.RandIntBetween(simState.Rand, int(startTime)+1, int(startTime+(60*60*12))))
 			}
 
-			bva := vestingtypes.NewBaseVestingAccount(&bacc, initialVesting, endTime)
+			bva := vestingtypes.NewBaseVestingAccount(bacc, initialVesting, endTime)
 
 			if simState.Rand.Intn(100) < 50 {
 				gacc = vestingtypes.NewContinuousVestingAccountRaw(bva, startTime)