+++ x/bank/types/codec.go
@@ -1,0 +1,28 @@
+package types
+
+import (
+	"github.com/cosmos/cosmos-sdk/codec"
+)
+
+// Register concrete types on codec codec
+func RegisterCodec(cdc *codec.Codec) {
+	cdc.RegisterConcrete(MsgSend{}, "cosmos-sdk/MsgSend", nil)
+	cdc.RegisterConcrete(MsgMultiSend{}, "cosmos-sdk/MsgMultiSend", nil)
+}
+
+var (
+	amino = codec.New()
+
+	// ModuleCdc references the global x/staking module codec. Note, the codec should
+	// ONLY be used in certain instances of tests and for JSON encoding as Amino is
+	// still used for that purpose.
+	//
+	// The actual codec used for serialization should be provided to x/staking and
+	// defined at the application level.
+	ModuleCdc = codec.NewHybridCodec(amino)
+)
+
+func init() {
+	RegisterCodec(amino)
+	amino.Seal()
+}